Esqueleto da tela de pagina inicial feito, codigo reorganizado para nao ficar monolítico
