+++ banco_app/pubspec.yaml
@@ -31,7 +31,7 @@ dependencies:
   flutter:
     sdk: flutter
   http: ^1.4.0
-  
+  google_fonts: ^6.2.1
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
